Move items into a shared folder
