+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/index.ts
@@ -3,13 +3,13 @@ const cors = require('cors');
 import 'express-async-errors';
 
 import { StudentController } from "./modules/students/student.controller";
-import { errorHandler } from './errorHandler';
+import { errorHandler } from './shared/errorHandler';
 import { ClassesController } from "./modules/classes/classes.controller";
 import { AssignmentsController } from "./modules/assignments/assignments.controller";
 import { AssignmentsService } from "./modules/assignments/assignments.service";
 import { ClassesService } from "./modules/classes/classes.service";
 import { StudentService } from "./modules/students/student.service";
-import { Database, prisma } from "./database";
+import { Database, prisma } from "./shared/database";
 
 const db = new Database(prisma);
 const studentService = new StudentService(db);
