+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/modules/assignments/assignments.controller.ts
@@ -1,5 +1,5 @@
 import express, { Request, Response, NextFunction } from "express";
-import { ErrorHandler } from "../../errorHandler";
+import { ErrorHandler } from "../../shared/errorHandler";
 import { parseForResponse } from "../../index";
 import { AssignmentsService } from "./assignments.service";
 import {
