+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/modules/assignments/assignments.service.ts
@@ -1,5 +1,5 @@
-import { Database } from "../../database";
-import { StudentAssignmentNotFoundException, StudentNotFoundException } from "../../exceptions";
+import { Database } from "../../shared/database";
+import { StudentAssignmentNotFoundException, StudentNotFoundException } from "../../shared/exceptions";
 import {
   AssignAssignmentToStudentDto,
   CreateClassAssignmentDto,
