+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/modules/classes/classes.service.ts
@@ -1,5 +1,5 @@
-import { Database, prisma } from "../../database";
-import { ClassNotFoundException, StudentAlreadyEnrolledException, StudentNotFoundException } from "../../exceptions";
+import { Database, prisma } from "../../shared/database";
+import { ClassNotFoundException, StudentAlreadyEnrolledException, StudentNotFoundException } from "../../shared/exceptions";
 import { CreateClassDto, EnrollStudentToClassDto, GetClassAssignmentsDto, GetClassByIdDto } from "./classes.dto";
 
 export class ClassesService {
