+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/modules/students/student.controller.ts
@@ -1,6 +1,6 @@
 import express, { Response, Request, NextFunction } from "express";
 
-import { ErrorHandler } from "../../errorHandler";
+import { ErrorHandler } from "../../shared/errorHandler";
 import { parseForResponse } from "../../index";
 import { StudentService } from "./student.service";
 import {
