+++ ThePhasesOfCraftship/2_best_practice_first/strategicDesignPart1/exercises/1_RefactoringTo4Tiers/begin/src/modules/students/student.dto.ts
@@ -1,5 +1,5 @@
 import { isMissingKeys, isUUID } from "../../index";
-import { InvalidRequestBodyException } from "../../exceptions";
+import { InvalidRequestBodyException } from "../../shared/exceptions";
 
 export class CreateStudentDto {
   private constructor(
